auth/requirements.txt
fastapi
uvicorn
pyjwt
cryptography
sqlalchemy